+++ packages/next-common/components/people/index.jsx
@@ -1,7 +1,7 @@
 import { useChainSettings } from "next-common/context/chain";
 import BaseLayout from "next-common/components/layout/baseLayout";
 import useRealAddress from "next-common/utils/hooks/useRealAddress";
-import NoWalletConnected from "next-common/components/assets/noWalletConnected";
+import NoWalletConnected from "next-common/components/assethubMigrationAssets/noWalletConnected";
 import { AccountImpl } from "next-common/components/layout/AccountLayout";
 import usePeopleOverviewTabs from "./overview/hooks/usePeopleOverviewTabs";
 import Tabs from "next-common/components/tabs";