+++ packages/next-common/utils/constants.js
@@ -37,6 +37,7 @@ import laos from "./consts/settings/laos";
 import paseoPeople from "./consts/settings/paseoPeople";
 import westendPeople from "./consts/settings/westendPeople";
 import hyperBridge from "./consts/settings/hyperBridge";
+import acurast from "./consts/settings/acurast";
 
 const optionalNodes =
   process.env.NEXT_PUBLIC_DEVELOPMENT === "true" ? [development] : [];
@@ -49,6 +50,7 @@ export const nodes = [
   // kusamaAssetHub,
   kusamaPeople,
   acala,
+  acurast,
   ajuna,
   astar,
   basilisk,
@@ -132,7 +134,7 @@ export const CACHE_KEY = {
   ambassadorDemotionExpireRemindVisible:
     "ambassador-demotion-expire-remind-visible",
   extensionUpdateMetadata: "extensionUpdateMetadata",
-  assetHubPromptVisible: "asset-hub-management-prompt-visible",
+  assetsPromptVisible: "assets-management-prompt-visible",
   multisigManagementPromptVisible: "multisig-management-prompt-visible",
   requestJudgementPrompt: "request-judgement-prompt",
   walletConnectSession: "walletconnect-session",
@@ -149,7 +151,7 @@ export const ADDRESS_CACHE_KEYS = [
   CACHE_KEY.ambassadorDemotionExpiredVisible,
   CACHE_KEY.ambassadorDemotionExpireRemindVisible,
   CACHE_KEY.extensionUpdateMetadata,
-  CACHE_KEY.assetHubPromptVisible,
+  CACHE_KEY.assetsPromptVisible,
   CACHE_KEY.multisigManagementPromptVisible,
   CACHE_KEY.requestJudgementPrompt,
   CACHE_KEY.setIdentityPromptVisible,